sales_api/pyproject.toml
[project]
name = "sales"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.9",
    "psycopg2-binary>=2.9.12",
]
